move some code to separate modules, work on adding sample playback
