+++ xmodits-gui/Cargo.toml
@@ -2,7 +2,7 @@
 name = "xmodits-gui"
 authors = ["B0ney <https://github.com/B0ney>"]
 description = "XMODITS graphical application"
-version = "0.10.0"
+version = "0.11.0"
 edition = "2021"
 build = "build.rs"
 license = "GPLv3"
@@ -14,7 +14,7 @@ winres = "0.1.12"
 
 [dependencies.xmodits-lib]
 git = "https://github.com/B0ney/xmodits-lib.git"
-rev = "65deb56"
+rev = "17a55e6"
 features = ["serde"]
 
 [dependencies.tokio]
@@ -69,6 +69,7 @@ cosmic-time = "0.1.2"
 iced_aw = "0.4.1"
 iced_gif = "0.3.0" 
 iced_lazy = "0.6.1"
+cpal = "0.15.2"
 
 [target.'cfg(windows)'.dependencies]
 winapi = { version = "0.3.9", features = ["shellapi", "winuser"] }
